+++ capnp-rpc/examples/calculator/server.rs
@@ -166,8 +166,7 @@ impl calculator::Server for CalculatorImpl {
     {
         Promise::from_future(async move {
             let v = evaluate_impl(params.get()?.get_expression()?, None).await?;
-            results.get().set_value(
-                calculator::value::ToClient::new(ValueImpl::new(v)).into_client::<::capnp_rpc::Server>());
+            results.get().set_value(capnp_rpc::new_client(ValueImpl::new(v)));
             Ok(())
         })
     }
@@ -177,10 +176,9 @@ impl calculator::Server for CalculatorImpl {
                     -> Promise<(), Error>
     {
         results.get().set_func(
-            calculator::function::ToClient::new(
+            capnp_rpc::new_client(
                 pry!(FunctionImpl::new(pry!(params.get()).get_param_count() as u32,
-                                       pry!(pry!(params.get()).get_body()))))
-                .into_client::<::capnp_rpc::Server>());
+                                       pry!(pry!(params.get()).get_body())))));
         Promise::ok(())
     }
     fn get_operator(&mut self,
@@ -189,8 +187,7 @@ impl calculator::Server for CalculatorImpl {
                     -> Promise<(), Error>
     {
         let op = pry!(pry!(params.get()).get_op());
-        results.get().set_func(
-            calculator::function::ToClient::new(OperatorImpl {op : op}).into_client::<::capnp_rpc::Server>());
+        results.get().set_func(capnp_rpc::new_client(OperatorImpl {op : op}));
         Promise::ok(())
     }
 }
@@ -207,8 +204,7 @@ pub async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     tokio::task::LocalSet::new().run_until(async move {
         let mut listener = tokio::net::TcpListener::bind(&addr).await?;
-        let calc =
-            calculator::ToClient::new(CalculatorImpl).into_client::<::capnp_rpc::Server>();
+        let calc: calculator::Client = capnp_rpc::new_client(CalculatorImpl);
 
         loop {
             let (stream, _) = listener.accept().await?;